requirements.txt
requests
django
pytest
djangorestframework

django-filter